Remove `WithNoVersionSchemaForRawSQL` option
This was a Xata specific option that was never exposed to users of the
CLI. If we need it in future, we can reintroduce it but there is no need
for it at the moment.

This change simplifies the code and removes the unnecessary complexity
of handling raw SQL migrations differently in various places.