Add new subcommand to update migrations that contain breaking changes (#855)
This PR adds a new subcommand called `update`. It updates migrations
that contain breaking changes in the selected folder.

```
Usage:
  pgroll update <directory> [flags]

Examples:
update ./migrations

Flags:
  -h, --help    help for update
  -j, --json    Output migration file in JSON format instead of YAML
```

### Dev notes

The new `migrations.FileUpdater` follows the chain of responsibility
design pattern. In the `FileUpdater` you can register new updater
functions (`migrations.updaterFn`) for any pgroll migration. The file
updater iterates over all operations in a migration and run the
registered updater functions in order.

The ordering of the updater functions is important. Please always add a
new updater function to the end of the list.

Closes #770

---------

Co-authored-by: Andrew Farries <andyrb@gmail.com>