Construct version schema name in `Roll.StartDDLOperations` without DB lookup (#912)
Change how the `Roll.StartDDLOperations` constructs the name of the
version schema that will be created by the migration so that it doesn't
require a lookup in the `pgroll.migrations` table.

There is no need to look it up in the state db; it's determied by
the`VersionSchema` of the migration being started and the schema name
it's being applied to.

Related to https://github.com/xataio/pgroll/issues/872