Always remove version schema on rollback and complete (#920)
Always clean up version schema on migation completion and rollback,
regardless of the value of the `disableVersionSchemas` flag.

It doesn't make sense to keep old version schema around after migration
completion/rollback as the views in the schema won't work after the
underlying tables are contracted.

The option makes sense only in the context of **creating** version
schema on migration start.

Part of #872