Simplify `previous_migration` function
Get rid of the recursive CTE in the `previous_migration` function - we
only need to find the parent of the latest migration, which can be done
with a simple query.

The need for recursion was removed because this function used to find
the previous migration with an existent version schema, but that
functionality is now in the `previous_version` function, allowing
`previous_migration` to be simplified.