Return `backfill.Task` from `Start` phase operations (#924)
From now on `Start` functions of each operations return a
`backfill.Task` instance. This struct contains information required by
the backfill process. The tasks are collected in a single
`backfill.Job`.
This aggregates all relevant information for the backfill process: list
of tables to backfill and all the triggers to load.