Support altering primary key columns that are not included in FK constraints (#930)
In the past `pgroll` could not preserve primary key constraint on
altered columns. Instead the primary key constraint was down-graded to a
unique, not null constraint. Now the PK constraint is not removed during
completing the migration.

Futhermore, if the table had rows, `pgroll` could not backfill rows.
`pgroll` "updates" the primary keys of the backfilled table to trigger
backfilling existing columns. Previously, finding the identity columns
returned the duplicated column `__pgroll_new_id` instead of the correct
`id`. In this case we need the current PK, so we can copy the correct
values to the new column.

I opened follow-up issue to support altering mutiple columns of the same
composite key in the same migration:
https://github.com/xataio/pgroll/pull/930

Support for alter PK columns that are part of FK constraints is in a
follow-up PR: https://github.com/xataio/pgroll/pull/935

Closes #888