Add trigger configuration to `backfill.Task` (#931)
This PR introduces consolidates backfilling of the same column into a
single trigger. I am merging all triggers for a table into a single
trigger in a follow-up PR.

I added a new data structure `OperationTrigger` that contains the
trigger configuration of an operation. Trigger consolidation happens
when a backfill task is added to the job. `triggerConfig` stayed the
consolidated configuration storage for the trigger template.

In a later PR I am removing `latestSchema` from the parameter list of
`Start`, as it is no longer necessary. We can pass it from `Roll` to the
`backfill.Job`: https://github.com/xataio/pgroll/pull/934