Return list of `DBAction` from `Start` (#944)
This PR returns adds a new return value to `Start` to return the list of
`DBActions`. However, returning 3 values is too many, so in a follow-up
PR I am combining the list of `DBActions` and `backfill.Task` into the
same data structure. The name is pending, my ideas so far `StartStep` or
`StartOperation`. If you have a better name, I am open to using them.

New data structure in a follow-up PR:
```golang
type StartOperation struct {
    Actions []DBAction
    Task *backfill.Task
}
```

---------

Co-authored-by: Andrew Farries <andyrb@gmail.com>