Recreate event trigger first in `init.sql`
Move event trigger creation to be the first thing that happens (after
`pgroll` schema creation) in `init.sql`.

Previous versions of `pgroll` used an event trigger that was
incompatible with the latest version, causing errors on the upgrade path
to `pgroll 0.14`. By moving event trigger creation to happen first
during `init`, we ensure that the old event trigger installed by an
older version of `pgroll` can't fire during the `init` of the new
version.