Remove the `--with-prefixes` flag from the `pull` command (#964)
Remove the `--with-prefixes` flag from the `pull` command.

The `--with-prefixes` flag was a workaround for inferred migrations not
being named in such a way that they would appear in the correct order
when pulled. As of https://github.com/xataio/pgroll/pull/899, inferred
migrations are named correctly so that `pgroll pull` will put all
migrations, including inferred migrations, in the correct order on disk.