Add `--expect-one` flag to `migrate` command (#1005)
Add an `--expect-one` flag to the `pgroll migrate` command to allow the
command to fail if there is more than one outstanding migration.

From the documentation added in this PR:

--- 

By default, `pgroll migrate` will apply all unapplied migrations.
However, it may sometimes be desirable to only apply a single migration
to ensure that an existing version schema is not removed by a sequence
of migrations. In this case, running:

```
$ pgroll migrate examples/ --expect-one
```

will cause the command to fail if more than one unapplied migration is
detected.