Fix `create_index` operation column ordering (#1009)
Change the `create_index` operation to take the columns on which the
index should be defined as an list rather than a map.

This ensures that the ordering of the columns in in the index is fixed
to the same order as they appear in the migration file.

**old format**

```yaml
# foo
operations:
  - create_index:
      name: some-index
      table: tickets
      columns:
        column1: {}
        column2: {}
```

**new format**

```yaml
operations:
- create_index:
    name: some-index
    table: tickets
    columns:
    - column: column1
      # options
    - column: column2
      # options
```

The changes in this PR are taken from
https://github.com/reve-ai/pgroll/pull/5 🙏

⚠️ This is a breaking change to the migration format ⚠️ 

We also modify the `pgroll update` command in this PR so that operations
in the old format can be migrated to the new format (although for
multi-column indexes the user will still have to ensure that the column
order in the updated migration is what they expected).

Fixes #1001